requirements.txt
fastapi
uvicorn
python-dotenv
pandas
scikit-learn
numpy
requests
schedule

openai